third_party/blink/web_tests/external/wpt/css/selectors/focus-visible-025-expected.txt

This is a testharness.js-based test.
PASS ":focus-visible" should be a valid selector
FAIL :focus-visible matches after accesskey on DIV after previous mouse focus assert_equals: outline-style for DIV#targetA should be solid expected "solid" but got "none"
FAIL :focus-visible matches after accesskey on BUTTON after previous mouse focus assert_equals: outline-style for DIV#initial should be none expected "none" but got "solid"
Harness: the test ran to completion.
